Move ServiceFileNames to lib so that generators can use it
