+++ lib/src/main/scala/generator/ServiceFileNames.scala
@@ -1,6 +1,7 @@
-package lib
+package generator
 
 import com.bryzek.apidoc.generator.v0.models.File
+import lib.Text
 
 object ServiceFileNames {
